refactor : Footer, Nav, SearchBar를 src의 common폴더로 분리
