+++ src/common/AppFooter.js
