+++ src/common/SearchBar.css
