+++ src/common/SearchBar.js
